feat: add remote browser PDF export via shared Playwright renderer
Move the Playwright-based browser PDF renderer out of the serverless
package into core/utils and leave a thin compatibility wrapper in the
serverless module so existing imports keep working while the Chromium
capture logic lives in one place.

Add Server.export_report_as_browser_pdf() to the remote server helper.
The new server-side path stages an offline HTML bundle through the
existing export_report_as_html() flow, validates Playwright options
before paying the download/render cost, and then prints that staged
bundle with the same offline Chromium technique that PR #517 added for
serverless exports.

Expose the remote capability through Report.export_browser_pdf() so the
higher-level report API can forward browser-PDF requests without callers
having to reach into serverobj directly.

Add focused coverage for:
- the shared Playwright renderer import path and existing renderer
  behavior
- remote browser PDF staging, option forwarding, and renderer failures
- Report.export_browser_pdf() forwarding and failure handling without a
  Docker-backed fixture

Refresh uv.lock after running uv sync --upgrade --all-extras and
uv lock --upgrade so the repo's security-scan expectations stay aligned
with the committed dependency graph.

Verification:
- make check
- make smoketest
- uv run pytest tests/serverless/test_pdf_renderer.py tests/test_report_remote_server.py::test_export_browser_pdf_stages_html_then_renders tests/test_report_remote_server.py::test_export_browser_pdf_requires_file_name tests/test_report_remote_server.py::test_export_browser_pdf_wraps_renderer_failures tests/test_report.py::test_export_browser_pdf_forwards_options tests/test_report.py::test_export_browser_pdf_returns_false_on_failure -q
